// backend/src/domain/enums/cuisine-type.enum.ts
export enum CuisineType {
  American = 'american',
  Barbecue = 'barbecue',
  Breakfast = 'breakfast',
  British = 'british',
  Cafe = 'cafe',
  Chinese = 'chinese',
  French = 'french',
  Greek = 'greek',
  Indian = 'indian',
  Italian = 'italian',
  Japanese = 'japanese',
  Mediterranean = 'mediterranean',
  Mexican = 'mexican',
  MiddleEastern = 'middle_eastern',
  Thai = 'thai',
  Vietnamese = 'vietnamese',
}